refactor(vfs): deduplicate path utilities into shared path-utils.ts (#752)
- Create lib/vfs/path-utils.ts with joinPath, basename, dirname, isAbsolutePath
- Remove duplicate implementations from electron-vfs.ts and web-vfs.ts
- Both VFS implementations now import from the shared module

Closes #752
